Implement evalToPolynomial further
We require multivariate polynomials and thus should use MvPolynomial.
Multivariate polynomials are required because the complexities we
measure are a function of the number of variables.

CommSemiring is required because MvPolynomial requires CommSemiRing.
CommRing is required because semirings are rings dropping the
requirement that each element must have an additive inverse.