Add basics for formulas (#3)
* Add Formula and size

They are still buggy.

* Lots of fixes

* Fix definition of size

The issues where mostly syntactical.

* Add a few examples for Formula and size

* Replace checks with definitions

* Prove examples for size

* Fix build-project workflow

The project was renamed to AlgebraicComplexity.

* Add definition for depth

* Delete project folder

The files were moved to AlgebraicComplexity.

* Add more examples

* Add examples for Formula.Neg

* some fixes

* Use new notation for examples

* Add draft version of evalToPolynomial

* Coerce constants to polynomial and add examples

* Ignore Test.lean

I want to use this file for testing stuff in Lean.

* Implement evalToPolynomial further

We require multivariate polynomials and thus should use MvPolynomial.
Multivariate polynomials are required because the complexities we
measure are a function of the number of variables.

CommSemiring is required because MvPolynomial requires CommSemiRing.
CommRing is required because semirings are rings dropping the
requirement that each element must have an additive inverse.

* Start fixing the examples for evalToPolynomial

* Switch CommSemiring back to Ring

I works also with Ring. I do not know what I did wrong earlier.

* Introduce n for Fin n

* Push fixes

* Fix example for evalToPolynomial

---------

Co-authored-by: Shreyas <shreyasss94@gmail.com>
Related to: https://github.com/ikelax/algebraic-complexity/issues/2